dynamically add event handlers in client (rm monitored)
Remove the `monitored` kwarg to `CytoscapeWidget`. Now the client code
automatically adds handlers that the kernel is listening to (still registered
using `CytoscapeWidget.on`) and dynamically adds listeners as the kernel
requests them.

Now the only public interface change from the master branch is the addition of
the `on` method for adding callbacks to browser events (remove callbacks with
`on(..., remove=True)`).